✨ feat: bank accounts

+++ src/modules/bankAccounts/services/DeleteBankAccountService.ts
@@ -3,7 +3,8 @@ import { inject, injectable } from 'tsyringe';
 import IBankAccountsRepository from '../repositories/IBankAccountsRepository';
 
 interface IRequest{
-    id: string
+    id: string;
+    user_id: string;
 }
 
 @injectable()
@@ -13,10 +14,15 @@ export default class DeleteBankAccountService {
         private bankAccountsRepository: IBankAccountsRepository
   ) {}
 
-  public async execute({ id }: IRequest): Promise<void> {
+  public async execute({ id, user_id }: IRequest): Promise<void> {
     const bankAccountWithId = await this.bankAccountsRepository.findById(id);
     if (!bankAccountWithId) {
       throw new AppError('No bank account with id');
+    }
+
+    // Não deixa qualquer user deletar a conta bancária
+    if (bankAccountWithId.user_id !== user_id) {
+      throw new AppError('Not authorized');
     }
     await this.bankAccountsRepository.delete(id);
   }
